backend/app/pyproject.toml
[tool.poetry]
name = "changachanga-backend"
version = "0.1.0"
description = ""
authors = ["Fokas Labs LTD <dann@fokas.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.13"}
sentry-sdk = {extras = ["fastapi"], version = "^1.22.2"}
redis = "^4.5.5"
celery = "^5.2.7"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
pytest = "^7.3.1"
pydantic = {extras = ["email"], version = "^1.10.7"}
passlib = "^1.7.4"
psycopg2 = "^2.9.6"
emails = "^0.6"
python-multipart = "^0.0.6"
alembic = "^1.10.4"
tenacity = "^8.2.2"
asyncpg = "^0.27.0"
bcrypt = "^4.0.1"
pyjwt = "^2.7.0"
gunicorn = "^20.1.0"
africastalking = "^1.2.5"
httpx = "^0.24.0"
boto = "^2.49.0"
boto3 = "^1.26.139"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
